# controller/openapi.yaml
openapi: 3.0.1
info:
  title: E-Commerce API
  version: 1.0.0
  description: API for use with e-commerce site.
paths:
  /users/login:
    summary: Log a user in or get the login page
    get:
      summary: 'Retrieve the Login Page'
      description: ''
      operationId: 'get_login_page'
      responses:
        '500':
          description: Default error response
        '200':
          description: Success
      tags:
        - Users
    post:
      summary: 'Login a User'
      description: ''
      operationId: 'get_user'
      requestBody:
        description: Object with username and password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        '500':
          description: Default error response
        '200':
          description: Success
      tags:
        - Users
  /users/logout:
    summary: Log out a user
    get:
      summary: 'Logout a User'
      description: 'logout_user'
      operationId: ''
      responses:
        '500':
          description: Default error response
        '200':
          description: Success
      tags:
        - Users
  /users/:
    summary: Get a user's information
    get:
      summary: 'Get a User'
      description: 'get_user'
      operationId: ''
      responses:
        '500':
          description: Default error response
        '200':
          content:
            application/json:
              examples:
                allProducts:
                  value: >-
                    {
                      "userInfo": {
                        "id": 123, 
                        "username": "Username",
                        "password": "Password123",
                        "email": "useremail@gmail.com",
                        "admin", false
                      }
                    };
          description: Success
      tags:
        - Users
  /users/register:
    summary: Register a user
    post:
      summary: 'Register a User'
      description: ''
      operationId: 'register_user'
      requestBody:
        description: Object with new username and password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
      responses:
        '500':
          description: Default error response
        '200':
          description: Success
      tags:
        - Users
  /users/username:
    summary: Change a user's username
    put:
      summary: 'Update a Username'
      description: ''
      operationId: 'update_username'
      requestBody:
        description: Object with new username
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Username'
      responses:
        '500':
          description: Default error response
        '200':
          description: Success
      tags:
        - Users
  /users/password:
    summary: Change a user's password
    put:
      summary: 'Update a password'
      description: ''
      operationId: 'update_password'
      requestBody:
        description: Object with new password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
      responses:
        '500':
          description: Default error response
        '200':
          description: Success
      tags:
        - Users
  /users/delete:
    summary: Delete a user
    delete:
      summary: 'Delete a User'
      description: ''
      operationId: 'delete_user'
      responses:
        '500':
          description: Default error response
        '200':
          description: Success
      tags:
        - Users
...
components:
  schemas:
    Login:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      xml:
        name: Login
    Register:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
      xml:
        name: Register
    Username:
      type: object
      properties:
        username:
          type: string
      xml:
        name: Username
    Password:
      type: object
      properties:
        password:
          type: string
      xml:
        name: Password
...